refactor(DL-16084): Addressed PlatOps comments for API routes
