+++ test/uk/gov/hmrc/entrydeclarationstore/controllers/api/DocumentationControllerSpec.scala
@@ -1,5 +1,5 @@
 /*
- * Copyright 2023 HM Revenue & Customs
+ * Copyright 2025 HM Revenue & Customs
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package uk.gov.hmrc.entrydeclarationstore.controllers
+package uk.gov.hmrc.entrydeclarationstore.controllers.api
 
 import controllers.Assets
 import org.scalatest.Assertion
